Create clip_shapefile.py
Problem: The shapefiles are required for each image tile.

Solution: Using the bounding box of each image tile, the entire shapefile of a city is clipped tile-wise.

Status: Ongoing. 
It does not perform automatic folder wise operation yet due to which the name of the city folder must be assigned for each run as ''uelzen'' presented above. Also shp_name.